Hid database accounts delete endpoint from production

+++ backend/database/db.py
@@ -20,6 +20,8 @@ async def lookup_user(identifier: str, db: Annotated[Client, Depends(get_db_conn
 
 @router.delete('/accounts/delete', status_code=status.HTTP_200_OK)
 async def delete_account(identifier: str, db:Annotated[Client, Depends(get_db_connection)]):
+    if os.environ.get('ENV') not in {'test', 'dev'}:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
     users_table = get_table_by_env('users')
     deletion_response = db.table(users_table).delete().or_(f'username.eq.{identifier},email.eq.{identifier}').execute()
     if not deletion_response.data:
